+++ test/setup/vitest.setup.js
@@ -15,6 +15,7 @@ import '../../utils/storage-manager.lock.js';
 import '../../utils/storage-manager.sync.js';
 import '../../utils/storage-manager.presets.js';
 import '../../utils/storage-manager.chatmap.js';
+import '../../utils/storage-manager.syncnow.js';
 import '../../content/censor-reply-restore.markdown.js';
 import '../../content/censor-reply-restore.dom.js';
 import '../../content/censor-reply-restore.storage.js';